add tests, services_routers logic

+++ src/api/v1/department/services.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from fastapi import HTTPException
 
+from schemas.schemas import UserToken
 from utils.service import BaseService
 from utils.unit_of_work import transaction_mode
 
@@ -9,8 +10,14 @@ from utils.unit_of_work import transaction_mode
 class OrganizationService(BaseService):
     @transaction_mode
     async def create_department(
-        self, name: str, company_id: int, parent_id: Optional[int] = None
+        self,
+        name: str,
+        company_id: int,
+        current_user: UserToken,
+        parent_id: Optional[int] = None,
     ):
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         parent_path = None
         if parent_id:
             parent = await self.uow.department.get_by_id(parent_id)
@@ -25,15 +32,34 @@ class OrganizationService(BaseService):
         )
 
     @transaction_mode
-    async def get_descendants(self, department_id: int) -> list:
+    async def get_descendants(
+        self,
+        department_id: int,
+        current_user: UserToken,
+    ) -> list:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         return await self.uow.department.get_descendants(department_id)
 
     @transaction_mode
-    async def get_ancestors(self, department_id: int) -> list:
+    async def get_ancestors(
+        self,
+        department_id: int,
+        current_user: UserToken,
+    ) -> list:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         return await self.uow.department.get_ancestors(department_id)
 
     @transaction_mode
-    async def move_department(self, department_id: int, new_parent_id: int) -> dict:
+    async def move_department(
+        self,
+        department_id: int,
+        new_parent_id: int,
+        current_user: UserToken,
+    ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         new_parent = await self.uow.department.get_by_id(new_parent_id)
         if not new_parent:
             raise HTTPException(
@@ -47,8 +73,14 @@ class OrganizationService(BaseService):
 
     @transaction_mode
     async def update_department(
-        self, department_id: int, name: Optional[str], parent_id: Optional[int]
+        self,
+        department_id: int,
+        name: Optional[str],
+        parent_id: Optional[int],
+        current_user: UserToken,
     ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         updates = {"name": name, "parent_id": parent_id}
         updates = {k: v for k, v in updates.items() if v is not None}
 
@@ -59,7 +91,14 @@ class OrganizationService(BaseService):
         return {"message": "Department updated successfully."}
 
     @transaction_mode
-    async def delete_department(self, department_id: int) -> dict:
+    async def delete_department(
+        self,
+        department_id: int,
+        current_user: UserToken,
+    ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
+
         department = await self.uow.department.get_by_id(department_id)
         if not department:
             raise HTTPException(status_code=404, detail="Department not found")
@@ -69,8 +108,14 @@ class OrganizationService(BaseService):
 
     @transaction_mode
     async def create_position(
-        self, name: str, description: Optional[str], company_id: int
+        self,
+        name: str,
+        description: Optional[str],
+        company_id: int,
+        current_user: UserToken,
     ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         position_id = await self.uow.position.add_one_and_get_id(
             name=name, description=description, company_id=company_id
         )
@@ -78,8 +123,14 @@ class OrganizationService(BaseService):
 
     @transaction_mode
     async def update_position(
-        self, position_id: int, name: Optional[str], description: Optional[str]
+        self,
+        position_id: int,
+        name: Optional[str],
+        description: Optional[str],
+        current_user: UserToken,
     ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         updates = {"name": name, "description": description}
         updates = {k: v for k, v in updates.items() if v is not None}
 
@@ -90,15 +141,25 @@ class OrganizationService(BaseService):
         return {"message": "Position updated successfully."}
 
     @transaction_mode
-    async def delete_position(self, position_id: int) -> dict:
+    async def delete_position(
+        self,
+        position_id: int,
+        current_user: UserToken,
+    ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         await self.uow.position.delete_one_by_id(obj_id=position_id)
         return {"message": "Position deleted successfully."}
 
     @transaction_mode
     async def assign_position_to_department(
-        self, position_id: int, department_id: int
+        self,
+        position_id: int,
+        department_id: int,
+        current_user: UserToken,
     ) -> dict:
-
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         position = await self.uow.position.get_by_query_one_or_none(id=position_id)
         if not position:
             raise HTTPException(status_code=404, detail="Position not found.")
@@ -118,7 +179,14 @@ class OrganizationService(BaseService):
         }
 
     @transaction_mode
-    async def assign_position_to_user(self, position_id: int, user_id: int) -> dict:
+    async def assign_position_to_user(
+        self,
+        position_id: int,
+        user_id: int,
+        current_user: UserToken,
+    ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         position = await self.uow.position.get_by_id(position_id)
         if not position:
             raise HTTPException(status_code=404, detail="Position not found.")
@@ -138,7 +206,14 @@ class OrganizationService(BaseService):
         return {"message": "Position assigned to user successfully"}
 
     @transaction_mode
-    async def assign_manager(self, department_id: int, user_id: int) -> dict:
+    async def assign_manager(
+        self,
+        department_id: int,
+        user_id: int,
+        current_user: UserToken,
+    ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         department = await self.uow.department.get_by_id(department_id)
         if not department:
             raise HTTPException(status_code=404, detail="Department not found")
@@ -147,14 +222,26 @@ class OrganizationService(BaseService):
         return {"message": "Manager assigned successfully"}
 
     @transaction_mode
-    async def get_subordinates(self, user_id: int) -> list:
+    async def get_subordinates(
+        self,
+        user_id: int,
+        current_user: UserToken,
+    ) -> list:
+        if not current_user.is_admin and current_user.user_id != user_id:
+            raise HTTPException(status_code=403, detail="Permission denied")
         subordinates = await self.uow.user.get_all_subordinates(user_id)
         return [sub.dict() for sub in subordinates]
 
     @transaction_mode
     async def assign_role(
-        self, user_id: int, department_id: int, role_name: str
+        self,
+        user_id: int,
+        department_id: int,
+        role_name: str,
+        current_user: UserToken,
     ) -> dict:
+        if not current_user.is_admin:
+            raise HTTPException(status_code=403, detail="Permission denied")
         user = await self.uow.user.get_by_id(user_id)
         department = await self.uow.department.get_by_id(department_id)
 
@@ -167,7 +254,13 @@ class OrganizationService(BaseService):
         return {"message": "Role assigned successfully."}
 
     @transaction_mode
-    async def get_roles(self, user_id: int) -> list:
+    async def get_roles(
+        self,
+        user_id: int,
+        current_user: UserToken,
+    ) -> list:
+        if not current_user.is_admin and current_user.user_id != user_id:
+            raise HTTPException(status_code=403, detail="Permission denied")
         roles = await self.uow.role_assignment.get_by_query_all(user_id=user_id)
         return [
             {"department_id": role.department_id, "role_name": role.role_name}
